Mentor update #1
By convention, when returning an error value, other values are not used, and it's considered idiomatic to return the zero value in this situation. The error value makes using an invalid "out of band" value to signal an error redundant in Go.
Variable diff is defined earlier than necessary. It's common when writing Go to declare variables as late as possible before they are used. This tends to reduce mental overhead when reading the code, and sometimes also leads to simpler code, since the code might return before the variable is needed.